+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Handle incomplete medata in newsfeed (#652)
 - Fix container image build lock, to prevent ImageEntrypoint concurrency issues (#671)
+- Fix filters on datamanager list view (#681)
 
 
 ## [0.8.0] 2022-01-16